chore(wallet-utils): Extract utility functions to new file
