+++ suite-common/wallet-config/src/index.ts
@@ -1,2 +1,3 @@
 export * from './networksConfig';
 export * from './types';
+export * from './utils';
